chore: frontend add call chatbot api
